- Patch by Damien Tournoud.  Refactors the form manipulation logic to make it cleaner and to avoid the side-effects of having to rebuild forms.
Defined a custom form #type = 'mollom', that we simply form-alter in protected forms. The analysis/captcha logic is performed in the #process of that form. This will work in any form types (including multi-step ones) because process functions are not cached.  As an added benefit, Mollom fields are now self-contained in the 'mollom' form element, so that 'session-id' doesn't pollute the form anymore.